% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rao_chr20_25_rep.R
\docType{data}
\name{rao_chr20_25_rep}
\alias{rao_chr20_25_rep}
\title{Contact matrix from Rao 2014, chromosome 20, 25kb resolution}
\format{A data.frame with 3 columns and 2125980 rows}
\usage{
rao_chr20_25_rep
}
\value{
A data.frame
}
\description{
A sparse 3-column contact matrix
}
\keyword{datasets}
